refactor: move PM resolver to emuplat and relocate constants
Address PR review feedback:
- Move UnderhillPmTimerAssistResolver and UnderhillPmTimerAssist from
  worker.rs to emuplat/pm_timer_assist.rs following existing patterns
- Replace local PM_BASE/SYSTEM_IRQ_ACPI constants with
  chipset_resources::pm::DEFAULT_PM_PIO_BASE and DEFAULT_ACPI_IRQ
- Remove now-unused chipset dependency from underhill_core
