+++ openhcl/underhill_core/Cargo.toml
@@ -41,7 +41,6 @@ vmm_core_defs.workspace = true
 state_unit.workspace = true
 vm_loader.workspace = true
 vnc_worker_defs.workspace = true
-chipset.workspace = true
 chipset_device.workspace = true
 chipset_device_resources.workspace = true
 chipset_device_worker.workspace = true
