+++ openhcl/underhill_core/src/emuplat/mod.rs
@@ -7,6 +7,7 @@ pub mod i440bx_host_pci_bridge;
 pub mod local_clock;
 pub mod netvsp;
 pub mod non_volatile_store;
+pub mod pm_timer_assist;
 pub mod tpm;
 pub mod vga_proxy;
 pub mod watchdog;
