+++ openhcl/underhill_core/src/worker.rs
@@ -193,9 +193,9 @@ use watchdog_core::resources::StaticWatchdogPlatformResolver;
 use zerocopy::FromZeros;
 
 #[cfg(guest_arch = "x86_64")]
-pub(crate) const PM_BASE: u16 = 0x400;
+use chipset_resources::pm::DEFAULT_ACPI_IRQ;
 #[cfg(guest_arch = "x86_64")]
-pub(crate) const SYSTEM_IRQ_ACPI: u32 = 9;
+use chipset_resources::pm::DEFAULT_PM_PIO_BASE;
 
 pub const UNDERHILL_WORKER: WorkerId<UnderhillWorkerParameters> = WorkerId::new("UnderhillWorker");
 
@@ -2214,9 +2214,11 @@ async fn new_underhill_vm(
     resolver.add_resolver(vmm_core::platform_resolvers::IoApicRoutingResolver(
         partition.ioapic_routing(),
     ));
-    resolver.add_resolver(UnderhillPmTimerAssistResolver {
-        partition: Arc::downgrade(&partition),
-    });
+    resolver.add_resolver(
+        crate::emuplat::pm_timer_assist::UnderhillPmTimerAssistResolver {
+            partition: Arc::downgrade(&partition),
+        },
+    );
 
     let bounce_buffer_tracker = {
         let size = {
@@ -2417,8 +2419,8 @@ async fn new_underhill_vm(
                     with_pic: capabilities.with_pic,
                     with_pit: capabilities.with_pit,
                     with_psp: dps.general.psp_enabled,
-                    pm_base: PM_BASE,
-                    acpi_irq: SYSTEM_IRQ_ACPI,
+                    pm_base: DEFAULT_PM_PIO_BASE,
+                    acpi_irq: DEFAULT_ACPI_IRQ,
                 },
             };
 
@@ -3883,52 +3885,6 @@ async fn load_firmware(
         .context("failed to set initial registers")?;
 
     Ok(())
-}
-
-pub struct UnderhillPmTimerAssist {
-    pub partition: std::sync::Weak<UhPartition>,
-}
-
-impl chipset::pm::PmTimerAssist for UnderhillPmTimerAssist {
-    fn set(&self, port: Option<u16>) {
-        if let Some(partition) = self.partition.upgrade() {
-            if let Err(err) = partition.set_pm_timer_assist(port) {
-                tracing::warn!(
-                    CVM_ALLOWED,
-                    error = &err as &dyn std::error::Error,
-                    ?port,
-                    "failed to set PM timer assist"
-                );
-            }
-        }
-    }
-}
-
-/// Resolver for the PM timer assist platform resource in Underhill.
-pub struct UnderhillPmTimerAssistResolver {
-    partition: std::sync::Weak<UhPartition>,
-}
-
-impl
-    vm_resource::ResolveResource<
-        chipset_resources::pm::PmTimerAssistHandleKind,
-        vm_resource::PlatformResource,
-    > for UnderhillPmTimerAssistResolver
-{
-    type Output = chipset_resources::pm::ResolvedPmTimerAssist;
-    type Error = std::convert::Infallible;
-
-    fn resolve(
-        &self,
-        _resource: vm_resource::PlatformResource,
-        _input: (),
-    ) -> Result<Self::Output, Self::Error> {
-        Ok(chipset_resources::pm::ResolvedPmTimerAssist(Box::new(
-            UnderhillPmTimerAssist {
-                partition: self.partition.clone(),
-            },
-        )))
-    }
 }
 
 // Represents a stub MMIO device that handles unhandled MMIO accesses by
